# public/swagger/afund.io.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "All Apis",
        "termsOfService": "",
        "contact": {
            "email": ""
        }
    },
    "host": "188.166.99.59:8080",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/users/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login user",
                "description": "Login user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Email and password will be passed through body",
                        "required": true,
                        "type": "string",
                        "default": "{\"email\" : \"test@test.com\",\"password\": \"123456\"}"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": []
            }
        },
        "/users": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "New user registartion",
                "description": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "All fields value required",
                        "required": true,
                        "type": "string",
                        "default": "{\"email\": \"test@test.com\", \"firstName\": \"Test\", \"lastName\": \"Test\", \"password\" : \"123456\"}"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": []
            }
        },
        "/users/change-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change user password",
                "description": "Change user password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "default": "",
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "All fields value required",
                        "required": true,
                        "type": "string",
                        "default": "{\"password\" : \"123456\",\"oldPassword\" : \"123456\"}"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": []
            }
        },
        "/users/send-recovery-mail": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Send password recovery information to the registered email",
                "description": "Send password recovery information to the registered email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "All fields value required",
                        "required": true,
                        "type": "string",
                        "default": "{\"email\" : \"test@test.com\"}"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": []
            }
        },
        "/users/reset-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Reset user password using email token",
                "description": "Reset user password using email token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "All fields value required",
                        "required": true,
                        "type": "string",
                        "default": "{\"password\" : \"123456\",\"validationToken\" : \"9c9baaa52d540bc5568e675f70f83810f81c946d9830471675bfda51432d52dc\"}"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": []
            }
        },
        "/users/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user's profile",
                "description": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "default": "",
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": []
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update user's profile",
                "description": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "default": "",
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "All fields value required",
                        "required": true,
                        "type": "string",
                        "default": "{\"firstName\": \"Test\", \"lastName\": \"Test\", \"bittrexApiKey\" : \"\", \"bittrexApiSecret\" : \"\"}"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": []
            }
        },
        "/users/profile-image": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update user's profile image",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "default": "",
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "profileImage",
                        "description": "All fields value required",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": []
            }
        },
...
    }
}